backend/pyproject.toml
[tool.poetry]
name = "mlops-backend"
version = "0.1.0"
description = ""
authors = ["Daniil Solovev <dan.solovjew@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.1"
aiohttp = "^3.9.5"
boto3 = "^1.34.122"
catboost = "^1.2.5"
mlflow = "2.13.1"
ipython = "8.24.0"
shapely = "^2.0.4"
prometheus-fastapi-instrumentator = "^7.0.0"


python-logging-loki = "^0.3.1"
[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
black = "^24.4.2"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.black]
line_length = 120

[tool.ruff.lint]
select = ["E", "F", "B", "D", "SIM", "I"]
ignore = ["F401", "D211", "D213", "B008", "D203"]

[tool.mypy]
strict = true
warn_unused_ignores = true
disable_error_code = "import-untyped"